% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessoryfunctions.R
\name{celltype_cluster_map}
\alias{celltype_cluster_map}
\title{Map Celltypes to Clusters using single cell per-cell metadata data.frame}
\usage{
celltype_cluster_map(
  celltype_cluster,
  cell_min_num = 10,
  singlect_thres = 0.9,
  multict_lothres = 0.5
)
}
\arguments{
\item{celltype_cluster}{data.frame with two columns, celltype (factor) and cluster (factor)}

\item{cell_min_num}{integer, default 10, min number of cells for celltype to be considered among the cluster}

\item{singlect_thres}{numeric 0 to 1, default 0.9. cutoff to assign a single celltype below this assign combined or just "Mixed"}

\item{multict_lothres}{numeric 0 to 1, default 0.5, cutoff to assign multiple celltypes if none pass singlect_thres, below this just assign "Mixed"}
}
\value{
a vector of length equal to the number of clusters, with the mapping of celltype(s) for each cluster. Can be used with \code{plyr::mapvalues()} with from = levels(cluster) to assign a new column of metadata.
}
\description{
Map Celltypes to Clusters using single cell per-cell metadata data.frame
}
\examples{
\dontrun{
#get md from serat obj
md <- sobj@meta.data

#prep the input df
celltype_cluster <- md[,c('Celltype', 'seurat_clusters')]

#make sure both are factorized
celltype_cluster[,1] <- factor(celltype_cluster[,1])
celltype_cluster[,2] <- factor(celltype_cluster[,2])

# run it
celltype_cluster_mappedlevs <- celltype_cluster_map(celltype_cluster)

}
}
