% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{deseq_to_gsea}
\alias{deseq_to_gsea}
\title{Wrapper script for FGSEA applied to MSIGDB subcategories}
\usage{
deseq_to_gsea(
  deseqres,
  gene_identifier_type = "gene_symbol",
  pathways,
  outdir,
  pathway_padj_thres = 0.1,
  pathway_pval_thres = Inf,
  workernum = 1,
  verbose = F,
  pwaycats = c("HALLMARK", "GO_BP", "GO_MF", "GO_CC", "CP_REACTOME", "CP_KEGG",
    "TFT_GTRD", "TFT_TFT_Legacy"),
  filename_prefix = ""
)
}
\arguments{
\item{deseqres}{data.frame, minimally must have columns named "log2FoldChange", "pvalue", and a gene column (with column name denoted by \code{gene_identifier_type})}

\item{gene_identifier_type}{string, default "gene_symbol", column name of gene identifiers. must be one of 'gene_symbol', 'gene_name', 'ensembl_gene', 'entrez_gene'. if 'gene_name' is passed, it will create and use 'gene_symbol'}

\item{pathways}{data.frame of pathways, output of \code{\link[=preppathways_pathwayanalysis_crosscondition_module]{preppathways_pathwayanalysis_crosscondition_module()}}, recommend you run this once and store use the pathways the way you would for a reference genome. Minimally, this is a data.frame where each row is a gene and the pathwy it is part of, with the following column names: "gs_name", the pathways; "gs_subcat", the categories/MSIGDB sub-categories of the pathways; and a column matching the parameter \code{gene_identifier_type}, such as "gene_symbol" for the genes in each pathway}

\item{outdir}{string, path to output directory. will create a subdir called "GSEA_tables"}

\item{pathway_padj_thres}{numeric, default 0.1, alpha for adjusted pvalue threshold for gsea pathways. Set to Inf (without quotes) to ignore the padj filter.}

\item{pathway_pval_thres}{numeric, default Inf, alpha for nominal pvalue threshold for gsea pathways, normally not used. If set to any value other than Inf, then pathway_padj_thres will be set to Inf.}

\item{workernum}{integer, default 1, number of CPUs, parallelization occurs over the eight database categories, max is 8}

\item{verbose}{T/F, default F, verbosity}

\item{pwaycats}{string or character vector. List of pathway subcategories to run. it should match the column of \code{pathways$gs_subcat}. Default is: c("HALLMARK", "GO_BP", "GO_MF", "GO_CC", "CP_REACTOME", "CP_KEGG", "TFT_GTRD", "TFT_TFT_Legacy")}

\item{filename_prefix}{string, default is empty. An optional string to add to all saved filenames, useful to open multiple tables later in excel.}
}
\description{
Will run FGSEA over the following MSIGB subcategories: "HALLMARK" (technical a category but coded as a sub-category here), "GO_BP", "GO_MF", "GO_CC", "CP_REACTOME", "CP_KEGG", "TFT_GTRD", "TFT_TFT_Legacy". Makes extensive use of the fgsea package. \href{https://www.biorxiv.org/content/10.1101/060012v3}{"Fast Gene Set Enrichment Analysis", Korotkevich et al bioRxiv 2021}. Ideally, GSEA should be used with the full result list.
}
