% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gseares_dotplot}
\alias{gseares_dotplot}
\title{Make a dotplot from fgsea result dataframe}
\usage{
gseares_dotplot(
  gseares,
  n_up = 5,
  n_dn = 5,
  line_char_width = 35,
  up_color = "red",
  midpoint_color = "white",
  dn_color = "steelblue",
  dotplot_fontsize = 5,
  include_pways
)
}
\arguments{
\item{gseares}{fgsea result data.frame}

\item{n_up}{integer, default 5, max number of upregulated pathways to plot (positive NES)}

\item{n_dn}{integer, default 5, max number of dnregulated pathways to plot (positive NES)}

\item{line_char_width}{integer, default 35, max character width per line of pathway y axis labels, newline added after words, see \code{\link[stringr:str_wrap]{stringr::str_wrap()}}}

\item{up_color}{color string name ort code, default 'red'}

\item{midpoint_color}{color string name ort code, default 'white'}

\item{dn_color}{color string name ort code, default 'steelblue'}

\item{dotplot_fontsize}{numeric, size of y axis pathway label font}

\item{include_pways}{vector of characters; if included, will always look for and include these even if not in n_up or n_dn}
}
\value{
a ggplot dotplot
}
\description{
Make a dotplot from fgsea result dataframe
}
