% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gseares_dotplot_listwrap}
\alias{gseares_dotplot_listwrap}
\title{Make a list of dotplots from a list of fgsea dataframes}
\usage{
gseares_dotplot_listwrap(
  gseareslist,
  outdir,
  plotprefix = "GSEA_Dotplot",
  pdfheight = 7,
  pdfwidth = 7,
  ...
)
}
\arguments{
\item{gseareslist}{output of \code{deseq_to_gsea()}}

\item{outdir}{string, path to output directory, will create a subdir called "GSEA_Category_Dotplots"}

\item{plotprefix}{string, default is 'GSEA_Dotplot', will name each of the category dotplot files with this prefix, ie "GSEA_Dotplot_HALLMARK.pdf". If changed, no need to put a "_" at the end, it is added automatically}

\item{pdfheight}{numeric, default 7}

\item{pdfwidth}{numeric, default 7}

\item{...}{other arguments passed to \code{\link[=gseares_dotplot]{gseares_dotplot()}}}
}
\value{
a list of ggplot dotplots
}
\description{
list wrapper around \code{\link[=gseares_dotplot]{gseares_dotplot()}}. see that function for details
}
