% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{plot_apear_cluster_pathway_dotplots}
\alias{plot_apear_cluster_pathway_dotplots}
\title{aPEAR cluster pathways dotplots}
\usage{
plot_apear_cluster_pathway_dotplots(
  apearoutlist,
  outdir,
  pdfheight = 7,
  pdfwidth = 7,
  ...
)
}
\arguments{
\item{apearoutlist}{the output of \code{\link[=gsea_to_aPEAR_clusters]{gsea_to_aPEAR_clusters()}}}

\item{outdir}{string, path to output directory, will create a subdir called "aPEAR/aPEAR_ClusterPathways_Dotplots"}

\item{pdfheight}{integer, default 7}

\item{pdfwidth}{integer, default 7}

\item{...}{other arguments passed to \code{\link[=gseares_dotplot]{gseares_dotplot()}}}
}
\description{
For each aPEAR cluster, make a dotplot of the pathways in that cluster
}
