% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{plot_apear_clusters}
\alias{plot_apear_clusters}
\title{Wrapper for aPEAR plotting}
\usage{
plot_apear_clusters(
  apearoutlist,
  outdir,
  date_plots = T,
  repelLabels = T,
  fontSize = 2.7,
  pdfheight = 7,
  pdfwidth = 10,
  ...
)
}
\arguments{
\item{apearoutlist}{the output of \code{\link[=gsea_to_aPEAR_clusters]{gsea_to_aPEAR_clusters()}}}

\item{outdir}{string, path to output directory, will create a subdir called "aPEAR/aPEAR_GraphPlots"}

\item{date_plots}{T/F, default T, whether to append datetime to apear graph plots, useful to remake plots in case they look weird (which is not uncommon)}

\item{repelLabels}{T/F, default T, adjust labels to avoid overlaps, highly imperfect}

\item{fontSize}{numeric, default 2.7, cluster label font size}

\item{pdfheight}{numeric, default 7}

\item{pdfwidth}{numeric, default 10}

\item{...}{other arguments passed to \code{\link[aPEAR:plotPathClusters]{aPEAR::plotPathClusters()}}}
}
\description{
A wrapper for plotting the aPEAR enrichment network. Sometimes looks weird, so it may be useful to run this a few times, will produce a different plot each time.
}
