% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessoryfunctions.R
\name{quickpdf}
\alias{quickpdf}
\title{Quick one-line wrapper for saving pdf}
\usage{
quickpdf(plotobject, filepath, pdfheight = 7, pdfwidth = 7, automakedir = F)
}
\arguments{
\item{plotobject}{any storable plot object than can be printed to pdf (ie ggplot)}

\item{filepath}{string, path to save to}

\item{pdfheight}{numeric, height of pdf}

\item{pdfwidth}{numeric, width of pdf}

\item{automakedir}{T/F, default F, whether to detect the folder path from \code{filepath} and recursively create the dir or not, can be dangerous and messy if not careful}
}
\description{
Quick one-line wrapper for saving pdf
}
