% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessoryfunctions.R
\name{quickpng}
\alias{quickpng}
\title{Quick one-line wrapper for saving png}
\usage{
quickpng(
  plotobject,
  filepath,
  plotheight = 7,
  plotwidth = 7,
  automakedir = F,
  pdfheight = NULL,
  pdfwidth = NULL,
  units = "in",
  res = 300
)
}
\arguments{
\item{plotobject}{any storable plot object than can be printed to pdf (ie ggplot)}

\item{filepath}{string, path to save to}

\item{plotheight}{numeric, height of plot file in units, default 7}

\item{plotwidth}{numeric, height of plot file in units, default 7}

\item{automakedir}{T/F, default F, whether to detect the folder path from \code{filepath} and recursively create the dir or not, can be dangerous and messy if not careful}

\item{pdfheight}{numeric, holdover from other plotting functions, will transfer to plotheight}

\item{pdfwidth}{numeric, holdover from other plotting functions, will transfer to plotwidth}

\item{units}{string, default 'in', see \code{?png}}

\item{res}{numeric, default 300, see \code{?png}}
}
\description{
Quick one-line wrapper for saving png
}
