% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_pipeline_wrapper.R
\name{gsea_apear_pipeline}
\alias{gsea_apear_pipeline}
\title{Full pipeline for DEGs to GSEA and aPEAR pathway clustering}
\usage{
gsea_apear_pipeline(
  deseqres,
  pathways,
  outdir,
  aPEAR_cluster_min_size = 3,
  min_pathway_gene_size = 3,
  num_input_sig_pathways_updn = 250,
  verbose = F,
  workernum = 1,
  filename_prefix = "",
  ...
)
}
\arguments{
\item{deseqres}{data.frame of differential expression, minimally must have columns named "log2FoldChange", "pvalue", and a gene column (with column name denoted by \code{gene_identifier_type})}

\item{pathways}{data.frame of pathways, output of \code{\link[=preppathways_pathwayanalysis_crosscondition_module]{preppathways_pathwayanalysis_crosscondition_module()}}, recommend you run this once and store use the pathways the way you would for a reference genome. Minimally, this is a data.frame where each row is a gene and the pathwy it is part of, with the following column names: "gs_name", the pathways; "gs_subcat", the categories/MSIGDB sub-categories of the pathways; and a column matching the parameter \code{gene_identifier_type}, such as "gene_symbol" for the genes in each pathway}

\item{outdir}{string, path to output directory. Will write result tables and plots to subdirectories inside of here.}

\item{aPEAR_cluster_min_size}{integer, default 3, minimum aPEAR cluster size after clustering the pathways}

\item{min_pathway_gene_size}{integer, default 3, minimum number of genes in pathway to be considered for clustering}

\item{num_input_sig_pathways_updn}{integer, default 250, max number of positive and negative NES pathways included, respectively. for example, when set to 250, 500 pathways maximum will be used, 250 positive and 250 negative}

\item{verbose}{T/F, default F}

\item{workernum}{integer, default 1, number of CPUs, parallelization occurs over the default database categories, so max is set to 8}

\item{filename_prefix}{string, default is empty. An optional string to add to all saved filenames, useful to open multiple tables later in excel.}

\item{...}{Other parameters passed on to \link{deseq_to_gsea}}
}
\value{
a list with elements containing outputs of all the modules
}
\description{
A wrapper pipeline function for GSEA pathway analysis and aPEAR clustering of enriched pathways, starting from differential expression (DE) summary statistics. Ideally, GSEA should be used with the full result list.
}
