% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessoryfunctions.R
\name{save_plotlist_suffix_listnames}
\alias{save_plotlist_suffix_listnames}
\title{Save a named list of plots to individual pdf files, one file per named list element}
\usage{
save_plotlist_suffix_listnames(
  plotfilename,
  plotlist,
  plotdir,
  pdfheight = NULL,
  pdfwidth = NULL,
  title_separator = "_",
  verbose = F
)
}
\arguments{
\item{plotfilename}{string, prefix that will be used to}

\item{plotlist}{the list to be plotted. must be named}

\item{plotdir}{the directory to plot to. will \emph{not} create a new dir (for safety)}

\item{pdfheight}{height of pdf}

\item{pdfwidth}{width of pdf}

\item{title_separator}{string, default '_', separator string between \code{plotfilename} and list element names in pdf file name}

\item{verbose}{T/F, default F, whether to report which element is being plotted during loop, useful to detect errors}
}
\value{
a list of files in the \code{plotdir}
}
\description{
given a named list of plots, save a pdf file for each list element
}
\examples{
\dontrun{
#set plotfilename, a string prefix
plotfilename <- 'ScatterPlots'

#set plotlist, a named list of plots, one pdf per list element;
# if any element is itself a list, the pdf will have pages for each sublist element
plotlist <- list(plot1 = plot(rnorm(100)), plot2 = plot(rnorm(100)))

#set the output folder
plotdir <- 'output_plots_will_go_here'

#title_separator; a string character that separates the 
# "plotfilename" prefix from the "plotlist" named list names
# here, the plots will be named like "ScatterPlots_plot1.pdf" and "ScatterPlots_plot2.pdf"
plotfilename = '_'

CVRCpipelines::save_plotlist_suffix_listnames(plotfilename,
                                              plotlist=plotlist,
                                              plotdir=plotdir)
}
}
